+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 01. :gift: Added class `dev.orne.http.client.BaseHttpServiceClient`
 01. :gift: Added interface `dev.orne.http.client.StatusDependentOperation`
 01. :gift: Added interface `dev.orne.http.client.StatusInitOperation`
+01. :gift: Added class `dev.orne.http.client.StatedHttpServiceClient`